channeldb: merge ControlTower with payments.go
